+++ CHANGELOG.md
@@ -23,10 +23,16 @@ This project adheres to
 - Updated Rust nightly version from `2024-06-13` to `2024-09-05` to match the
   version used in Mina proof-systems repository.
 
+### Added
+
+  Added `caml_fp_srs_get_lagrange_basis_ptr` and `caml_fq_srs_get_lagrange_basis_ptr` pointer functions to enable using a browser cache with `zkProgram.compile()`. https://github.com/o1-labs/o1js/pull/2404
+
 ### Fixed
 
 - Fixed a performance regression that occured when proving circuits.
   https://github.com/o1-labs/o1js/pull/2388
+
+- Fixed issue where `zkProgram.compile()` was not able to use a browser cache. Previously, the `get_lagrange_basis` functions were disabled on the web because they would hang indefinitely when cache writing was enabled. The issue was caused by lagrange basis functions returning types incompatible with web worker transfer. https://github.com/o1-labs/o1js/pull/2404
 
 ## [2.8.0](https://github.com/o1-labs/o1js/compare/045b1ab...4b1dccdd) - 2025-08-01